fix: propertie with wrong type

+++ lbaas/network_backend.go
@@ -31,9 +31,9 @@ type (
 	}
 
 	UpdateNetworkBackendRequest struct {
-		HealthCheckID                       *string  `json:"health_check_id,omitempty"`
-		PanicThreshold                      *float64 `json:"panic_threshold,omitempty"`
-		CloseConnectionsOnHostHealthFailure *bool    `json:"close_connections_on_host_health_failure,omitempty"`
+		HealthCheckID                       *string `json:"health_check_id,omitempty"`
+		PanicThreshold                      *int    `json:"panic_threshold,omitempty"`
+		CloseConnectionsOnHostHealthFailure *bool   `json:"close_connections_on_host_health_failure,omitempty"`
 	}
 
 	NetworkBackedTarget struct {
@@ -51,7 +51,7 @@ type (
 		Name                                string                  `json:"name"`
 		Description                         *string                 `json:"description,omitempty"`
 		BalanceAlgorithm                    BackendBalanceAlgorithm `json:"balance_algorithm"`
-		PanicThreshold                      *float64                `json:"panic_threshold,omitempty"`
+		PanicThreshold                      *int                    `json:"panic_threshold,omitempty"`
 		CloseConnectionsOnHostHealthFailure *bool                   `json:"close_connections_on_host_health_failure,omitempty"`
 		TargetsType                         BackendType             `json:"targets_type"`
 		Targets                             []NetworkBackedTarget   `json:"targets"`
